refactor version helpers
