+++ apps/studio/src/common/appdb/models/LicenseKey.ts
@@ -1,5 +1,6 @@
 import platformInfo from "@/common/platform_info";
-import { LicenseStatus, isVersionLessThanOrEqual } from "@/lib/license";
+import { LicenseStatus } from "@/lib/license";
+import { isVersionLessThanOrEqual } from "@/common/version";
 import { Column, Entity, Not } from "typeorm";
 import { ApplicationEntity } from "./application_entity";
 import _ from 'lodash';
