+++ apps/studio/src/lib/license.ts
@@ -1,13 +1,7 @@
 import { TransportLicenseKey } from "@/common/transport";
-import _ from "lodash";
+import { parseVersion, Version } from "@/common/version";
 
-export interface BaseVersion {
-  major: number;
-  minor: number;
-  patch: number;
-}
-
-export interface Version extends BaseVersion {
+export interface BksVersion extends Version {
   channel: 'stable' | 'beta' | 'alpha'
   channelRelease?: number
 }
@@ -19,29 +13,6 @@ export enum DevLicenseState {
   activePaidLicense,
   expiredLifetimeCoversThisVersion,
   expiredLifetimeCoversEarlierVersion,
-}
-
-/**
- * Usage:
- *
- * ```
- * const v1 = parseVersion('5.1.0')
- * console.log(v1.major, v1.minor, v.patch)
- * ```
- **/
-export function parseVersion(version: string) {
-  const versionTagRegex = /^(\d+)\.(\d+)\.(\d+)$/;
-  const match = versionTagRegex.exec(version) || [];
-  const [major, minor, patch] = _.tail(match).map((x) => parseInt(x));
-  return { major, minor, patch };
-}
-
-/** Check if version a is less than or equal to version b */
-export function isVersionLessThanOrEqual(a: BaseVersion, b: BaseVersion) {
-  if (a.major > b.major) return false;
-  if (a.minor > b.minor) return false;
-  if (a.patch > b.patch) return false;
-  return true;
 }
 
 export class LicenseStatus {
